+++ src/test/kotlin/io/github/ackeecz/danger/lint/detekt/DetektReportProcessorTest.kt
@@ -1,31 +1,32 @@
-package io.github.ackeecz.danger.detekt
+package io.github.ackeecz.danger.lint.detekt
 
+import io.github.ackeecz.danger.lint.FakeDangerContext
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.collections.shouldBeEmpty
 import io.kotest.matchers.collections.shouldHaveSize
 import io.kotest.matchers.shouldBe
 import java.io.File
 
-internal class DetektPluginTest : FunSpec({
+internal class DetektReportProcessorTest : FunSpec({
 
     test("should process two files in report") {
         val context = FakeDangerContext()
-        DetektPlugin.context = context
+        val underTest = DetektReportProcessor(context)
 
-        DetektPlugin.parseAndReport(
-            loadFile("detekt_result_two_files.xml"),
-        )
+        underTest.process(listOf(loadFile("detekt_result_two_files.xml")))
 
         context.warnings shouldHaveSize 6
     }
 
     test("should process two report files") {
         val context = FakeDangerContext()
-        DetektPlugin.context = context
+        val underTest = DetektReportProcessor(context)
 
-        DetektPlugin.parseAndReport(
-            loadFile("detekt_result_two_files.xml"),
-            loadFile("detekt_result_single_file.xml"),
+        underTest.process(
+            listOf(
+                loadFile("detekt_result_two_files.xml"),
+                loadFile("detekt_result_single_file.xml"),
+            )
         )
 
         context.warnings shouldHaveSize 8
@@ -33,33 +34,27 @@ internal class DetektPluginTest : FunSpec({
 
     test("should process one file in report") {
         val context = FakeDangerContext()
-        DetektPlugin.context = context
+        val underTest = DetektReportProcessor(context)
 
-        DetektPlugin.parseAndReport(
-            loadFile("detekt_result_single_file.xml")
-        )
+        underTest.process(listOf(loadFile("detekt_result_single_file.xml")))
 
         context.warnings shouldHaveSize 2
     }
 
     test("should process no file in report") {
         val context = FakeDangerContext()
-        DetektPlugin.context = context
+        val underTest = DetektReportProcessor(context)
 
-        DetektPlugin.parseAndReport(
-            loadFile("detekt_result_no_files.xml")
-        )
+        underTest.process(listOf(loadFile("detekt_result_no_files.xml")))
 
         context.warnings.shouldBeEmpty()
     }
 
     test("should have relative path in filename") {
         val context = FakeDangerContext()
-        DetektPlugin.context = context
+        val underTest = DetektReportProcessor(context)
 
-        DetektPlugin.parseAndReport(
-            loadFile("detekt_result_single_file.xml"),
-        )
+        underTest.process(listOf(loadFile("detekt_result_single_file.xml")))
 
         context.warnings.first().file shouldBe "features/recipelist/src/main/java/cz/ackee/sample/recipelist/presentation/RecipeListFragment.kt"
     }